chore: moved mssql mock file from root __mocks__ dir
fix: failing contract test and migrations error because automatically using mocked mssql
